requirements.txt
gradio
openai
python-dotenv
requests
googlemaps
folium